refactor: rename src/graph/ → src/exec/
The Graph IR (OpNode, OpType, Graph) was deleted in #287. The directory
name "graph" was misleading — its contents are imperative GraphExecutor
/ GEMM-kernel-registry / workspace / scratch code, not a graph. Rename
clarifies intent.

Mechanical change: file moves + include-path rewrite + CMake variable
rename (IMP_GRAPH_SOURCES → IMP_EXEC_SOURCES). make verify-fast green.

Phase 1 (soft PR) of docs/superpowers/specs/2026-05-20-architecture-refactor-roadmap-design.md
