+++ src/exec/executor.h
@@ -12,9 +12,9 @@
 #include "compute/gemm_cutlass_sm120.h"
 #include "compute/gemm_cutlass_mxfp4_sm120.h"
 #include "core/tensor.h"
-#include "graph/weight_handle.h"
-#include "graph/moe_workspace.h"
-#include "graph/quant_scratch.h"
+#include "exec/weight_handle.h"
+#include "exec/moe_workspace.h"
+#include "exec/quant_scratch.h"
 #include "runtime/storage_planner.h"
 #include <cuda_runtime.h>
 #include <cuda_fp16.h>
