+++ src/exec/executor_ffn.cu
@@ -1,8 +1,8 @@
-#include "graph/executor.h"
-#include "graph/executor_kernels.h"
-#include "graph/executor_gemv_helpers.h"
-#include "graph/executor_helpers.h"
-#include "graph/gemm_context.h"
+#include "exec/executor.h"
+#include "exec/executor_kernels.h"
+#include "exec/executor_gemv_helpers.h"
+#include "exec/executor_helpers.h"
+#include "exec/gemm_context.h"
 #include "compute/layernorm.h"
 #include "compute/gemm.h"
 #include "compute/gemm_q6k.h"
