+++ src/exec/executor_ssm_gdn.cu
@@ -1,7 +1,7 @@
-#include "graph/executor.h"
-#include "graph/executor_kernels.h"
-#include "graph/executor_debug.h"
-#include "graph/gemm_context.h"
+#include "exec/executor.h"
+#include "exec/executor_kernels.h"
+#include "exec/executor_debug.h"
+#include "exec/gemm_context.h"
 #include "compute/layernorm.h"
 #include "compute/activation.h"
 #include "compute/ssm.h"
