+++ src/exec/gemm_kernel_gguf.cu
@@ -1,11 +1,11 @@
-#include "graph/gemm_kernel_registry.h"
+#include "exec/gemm_kernel_registry.h"
 
 #include "compute/gemm.h"        // quantize_fp16_to_q8_1, block_q8_1
 #include "compute/ggml_mmvq.h"   // ggml_mmvq_q4k / q5k / q5_1 / q8_0
 #include "core/logging.h"
 #include "core/tensor.h"
-#include "graph/executor_kernels.h"  // is_dp4a_qtype, dispatch_dp4a_gemv, block_q8_1
-#include "graph/gemm_scratch.h"  // mmvq_scratch_get_or_grow (Slice 8.6 hoist)
+#include "exec/executor_kernels.h"  // is_dp4a_qtype, dispatch_dp4a_gemv, block_q8_1
+#include "exec/gemm_scratch.h"  // mmvq_scratch_get_or_grow (Slice 8.6 hoist)
 #include "runtime/config.h"
 
 #include <cuda_fp16.h>
