+++ src/exec/moe_workspace.cu
@@ -1,4 +1,4 @@
-#include "graph/moe_workspace.h"
+#include "exec/moe_workspace.h"
 #include "core/logging.h"
 #include <cuda_runtime.h>
