+++ src/exec/quant_scratch.h
