+++ src/exec/weight_handle.cu
@@ -1,4 +1,4 @@
-#include "graph/weight_handle.h"
+#include "exec/weight_handle.h"
 #include "memory/vram_allocator.h"
 #include "core/logging.h"
